cdk/requirements.txt
aws-cdk-lib==2.166.0
constructs>=10.0.0,<11.0.0
pillow==11.0.0
numpy==2.0.2
boto3==1.35.63
aws-cdk-core=1.204.0